pyproject.toml

[tool.poetry]
name = "deploy-ml-bento"
version = "0.1.0"
description = "Este projeto tem o objetivo de desenvolver o deploy de um modelo"
authors = ["Espedito Ferreira Alves"]
package-mode = false  # Desativa o empacotamento do projeto, utilizando o Poetry apenas para gerenciamento de dependências

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
bentoml = "^1.2.19"
matplotlib = "^3.9.1"
yellowbrick = "^1.5"
streamlit = "1.10.0"
nest-asyncio = "^1.6.0"
numpy = "^2.0.0"
pydantic = "^2.8.2"
altair = "4.1.0"

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"
pandas-stubs = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
ignore_missing_imports = true